# Optimize JSON parsing a bit