--- a/spec/std/json/pull_parser_spec.cr
+++ b/spec/std/json/pull_parser_spec.cr
@@ -92,8 +92,8 @@ class JSON::PullParser
   end
 end
 
-private def assert_pull_parse(string)
-  it "parses #{string}" do
+private def assert_pull_parse(string, file = __FILE__, line = __LINE__)
+  it "parses #{string}", file, line do
     parser = JSON::PullParser.new string
     parser.assert JSON.parse(string).raw
     parser.kind.should eq(JSON::PullParser::Kind::EOF)