--- a/src/json/lexer/io_based.cr
+++ b/src/json/lexer/io_based.cr
@@ -2,13 +2,13 @@
 class JSON::Lexer::IOBased < JSON::Lexer
   def initialize(@io : IO)
     super()
-    @current_char = @io.read_char || '\0'
+    @current_char = @io.read_byte.try(&.chr) || '\0'
   end
 
   private getter current_char
 
   private def next_char_no_column_increment
-    @current_char = @io.read_char || '\0'
+    @current_char = @io.read_byte.try(&.chr) || '\0'
   end
 
   private def consume_string