--- a/src/json/lexer/io_based.cr
+++ b/src/json/lexer/io_based.cr
@@ -12,7 +12,48 @@ class JSON::Lexer::IOBased < JSON::Lexer
   end
 
   private def consume_string
-    consume_string_with_buffer
+    peek = @io.peek
+    if !peek || peek.empty?
+      return consume_string_with_buffer
+    end
+
+    pos = 0
+
+    while true
+      if pos >= peek.size
+        # We don't have enough data in the peek buffer to create a string:
+        # default to the slow method
+        @column_number -= pos
+        return consume_string_with_buffer
+      end
+
+      char = peek[pos]
+      case char
+      when '\\'
+        # If we find an escape character, go to the slow method
+        @column_number -= pos
+        return consume_string_with_buffer
+      when '"'
+        @column_number += 1
+        @io.skip(pos + 1)
+        @current_char = @io.read_byte.try(&.chr) || '\0'
+        break
+      else
+        if 0 <= current_char.ord < 32
+          unexpected_char
+        else
+          pos += 1
+          @column_number += 1
+        end
+      end
+    end
+
+    @token.string_value =
+      if @expects_object_key
+        @string_pool.get(peek.to_unsafe, pos)
+      else
+        String.new(peek.to_unsafe, pos)
+      end
   end
 
   private def number_start