--- a/src/json/token.cr
+++ b/src/json/token.cr
@@ -19,7 +19,7 @@ class JSON::Token
   property string_value : String
 
   def int_value : Int64
-    raw_value.to_i64
+    @int_value || raw_value.to_i64
   rescue exc : ArgumentError
     raise ParseException.new(exc.message, line_number, column_number)
   end
@@ -32,7 +32,8 @@ class JSON::Token
 
   property line_number : Int32
   property column_number : Int32
-  property raw_value : String
+  setter raw_value : String
+  setter int_value : Int64?
 
   def initialize
     @kind = :EOF
@@ -40,6 +41,16 @@ class JSON::Token
     @column_number = 0
     @string_value = ""
     @raw_value = ""
+    @int_value = nil
+  end
+
+  def raw_value
+    case @kind
+    when .int?
+      @int_value.try(&.to_s) || @raw_value
+    else
+      @raw_value
+    end
   end
 
   def to_s(io : IO) : Nil
@@ -51,7 +62,11 @@ class JSON::Token
     when .true?
       io << "true"
     when .int?
-      raw_value.to_s(io)
+      if int_value = @int_value
+        int_value.to_s(io)
+      else
+        raw_value.to_s(io)
+      end
     when .float?
       raw_value.to_s(io)
     when .string?