I saw [this thread](https://forum.crystal-lang.org/t/performance-issues-with-the-json-parser/6678), so...

Two things here:
1. Read string/IO byte per byte instead of char per char
2. When trying to parse a string in the IO case, try to use the peek buffer if there's one

I used this Ruby/Crystal file to generate a big JSON:

```crystal
json = %({"foo": 1, "bar": 2, "hello": "this will have an escape in it\\\"oh well", "banana": true, "array": [1, 2, 3, "a string that is a bit long"], "hash": {"a": 1, "b": 2, "c": 3}})
file = "[" + ([json] * 100000).join(",") + "]"
File.write("json.json", file)
```

Here's the benchmark:

```crystal
require "json"
require "benchmark"

file_io = File.open("json.json")
file_string = File.read("json.json")

Benchmark.ips do |x|
  x.report("JSON.parse (string)") do
    JSON.parse(file_string)
  end
  x.report("JSON.parse (IO)") do
    file_io.pos = 0
    JSON.parse(file_io)
  end
end
```

Before:
```
JSON.parse (string)   8.41  (118.97ms) (± 0.86%)  117MB/op        fastest
    JSON.parse (IO)   3.64  (274.72ms) (± 1.99%)  117MB/op   2.31× slower
```

After:
```
JSON.parse (string)  11.98  ( 83.50ms) (± 7.70%)  117MB/op        fastest
    JSON.parse (IO)   9.06  (110.41ms) (± 6.26%)  117MB/op   1.32× slower
```

I think the difference between before and after will be bigger if there are more strings to parse.

A note for the forum thread: I tried parsing that same big file with Ruby 3.1 and Ruby was (slightly) slower: 16 seconds in Crystal vs. 19 seconds in Ruby. This is on a Mac. So I don't know why it was slower in Crystal for OP (maybe in Linux it's different?)

Regarding memory: Crystal requires 117MB to load that entire data into memory. But in Ruby it's the same. So I'm not sure how memory can be further optimized...

Please review carefully! I think tests for when the peek buffer is incomplete or unavailable might not exist right now. Feel free to continue working on top of this PR (pushing commits to this branch or creating another PR from this code).