@asterite 2024-03-16
I found another optimization that improved both speed of parsing and the memory it allocates.

[This PR](https://github.com/crystal-lang/crystal/pull/10447) made parsing slower by always allocating a string to parse integers and floats from. It also made it consume more memory. I was [concerned about it](https://github.com/crystal-lang/crystal/pull/10447#issuecomment-786753382) at that time exactly because of that, but I agreed that correctness is more important than performance. However, we can have both! For "small" integers (less than 19 digits, never floats) we can compute the int value and always know that we are doing it right.

Before the last commit:

```
JSON.parse (string)  10.76  ( 92.92ms) (± 8.00%)  125MB/op        fastest
    JSON.parse (IO)   7.51  (133.21ms) (± 5.99%)  125MB/op   1.43× slower
```

After the last commit:

```
JSON.parse (string)  12.09  ( 82.69ms) (± 1.28%)  105MB/op        fastest
    JSON.parse (IO)   8.06  (124.08ms) (± 2.37%)  105MB/op   1.50× slower
```

So a bit faster and 20MB less, so about 16% percent less in this case.

But the memory allocated here depends on the amount of numbers and how long they are. For example, using this file:

```ruby
json = "[1234234, 2982374, 3982734, 49827344, 592834, 65825, 723498, 82348, 9239847324, 1082348, 1123498, 122348, 132348, 142347, 152348, 16234283]"
file = "[" + ([json] * 10000).join(",") + "]"
File.write("json.json", file)
```

running the benchmark, before this PR:

```
JSON.parse (string)  95.86  ( 10.43ms) (± 1.44%)  13.9MB/op        fastest
    JSON.parse (IO)  42.03  ( 23.79ms) (± 3.17%)  13.9MB/op   2.28× slower
```

after this PR:

```
JSON.parse (string) 205.35  (  4.87ms) (± 1.11%)  9.06MB/op        fastest
    JSON.parse (IO)  76.83  ( 13.02ms) (± 1.18%)  9.06MB/op   2.67× slower
```

so 5 MB less than before, from a total of 14MB, that's about 30% less memory!