@philipp-kempgen 2024-03-17
> A note for the forum thread: I tried parsing that same big file with Ruby 3.1 and Ruby was (slightly) slower: 16 seconds in Crystal vs. 19 seconds in Ruby. This is on a Mac. So I don't know why it was slower in Crystal for OP

Let's make the JSON data contain some long strings. They are Base64-encoded in my case, but that doesn't matter.

`gen-json.rb`:
```rb
json = %|{"a_base64":#{("a" * 5000).inspect},"b_base64":#{("a" * 10000).inspect}}|
json = "[" + ([json] * 2500).join(",") + "]"
File.write("json.json", json)
```

`benchmark-pk.cr`:
```cr
require "json"
require "benchmark"

file_io = File.open("json.json")
file_string = File.read("json.json")

Benchmark.bm do |x|
  x.report("JSON.parse (string)") do
    JSON.parse(file_string)
  end
end
```

`benchmark-pk.rb`:
```rb
require "json"
require "benchmark"

file_io = File.open("json.json")
str = File.read("json.json")

Benchmark.bm(19) do |x|
  x.report("JSON.parse(str)") do
    JSON.parse(str)
  end
end
```

```sh
$ crystal build --release benchmark-pk.cr
$ /usr/bin/time -l ./benchmark-pk
                          user     system      total        real
JSON.parse (string)   0.174718   0.005743   0.180461 (  0.180668)
        0,20 real         0,18 user         0,02 sys
            94814208  maximum resident set size
                   0  average shared memory size
                   0  average unshared data size
                   0  average unshared stack size
                5877  page reclaims
                   0  page faults
                   0  swaps
                   0  block input operations
                   0  block output operations
                   0  messages sent
                   0  messages received
                   0  signals received
                   0  voluntary context switches
                  39  involuntary context switches
          1868924524  instructions retired
           591412605  cycles elapsed
            93389632  peak memory footprint
```

That's *before* your changes, on Apple aarch64/arm64.

```sh
$ /usr/bin/time -l ruby ./benchmark-pk.rb
                          user     system      total        real
JSON.parse(str)       0.063432   0.003553   0.066985 (  0.067092)
        0,12 real         0,10 user         0,01 sys
            90554368  maximum resident set size
                   0  average shared memory size
                   0  average unshared data size
                   0  average unshared stack size
                5642  page reclaims
                   0  page faults
                   0  swaps
                   0  block input operations
                   0  block output operations
                   0  messages sent
                   0  messages received
                   0  signals received
                   0  voluntary context switches
                  43  involuntary context switches
          1390183537  instructions retired
           367126927  cycles elapsed
            86574272  peak memory footprint
```

That's using Ruby 3.3.0.
The Ruby version is faster by a factor of 0.181/0.067 = 2.70.