@jzakiya 2024-03-17
Starting with Ruby 3.3 you can enable YJIT at runtime.
It would be nice to see the w/wo YJIT times for this.

I now put the following Ruby snippet at the top of Ruby files to
automatically enable YJIT when using Ruby >=3.3.

```
# Enable YJIT if using CRuby >= 3.3"
RubyVM::YJIT.enable if RUBY_ENGINE == 'ruby' and RUBY_VERSION.to_f >= 3.3
```