@philipp-kempgen 2024-03-17
> Starting with Ruby 3.3 you can enable YJIT at runtime. It would be nice to see the w/wo YJIT times for this.

I get the same times with or without YJIT.