@philipp-kempgen 2024-03-17
And just for the record, here's my benchmark in Ruby with [Oj](https://rubygems.org/gems/oj):

```rb
require "benchmark"
require "json"
require "oj"

file_io = File.open("json.json")
str = File.read("json.json")

Benchmark.bm(19) do |x|
  x.report("JSON.parse(str)") do
    JSON.parse(str)
  end
  x.report("Oj.load(str)") do
    Oj.load(str)
  end
end
```

```
                          user     system      total        real
JSON.parse(str)       0.063371   0.003690   0.067061 (  0.067182)
Oj.load(str)          0.028192   0.003248   0.031440 (  0.031523)
```

i.e. Oj is faster than the Crystal version by a factor of 0.181/0.032 = 5.66.